+++ src/scripts/engine.ts
@@ -3,9 +3,9 @@ import * as PIXI from "pixi.js"
 import useEngine from "../composeables/use-engine"
 import { EntityMap } from "./entities/create-entity"
 import { Entity } from "./entities/entity"
-import { demoWorld, secondWorld, startWorld as startWorldRT } from "./rt/util/templates-world"
+import { demoWorld, secondWorld, startWorld as startWorldRT } from "./module-data/rt/util/templates-world"
+import { startWorld as startWorldTB } from "./module-data/tb/util/template-world"
 import { RenderSystem } from "./systems/render"
-import { startWorld as startWorldTB } from "./tb/util/template-world"
 import { CreateWorld, WorldMap } from "./util/create-world"
 import { LocalStorageManager } from "./util/local-storage-manager"
 import { SaveManager } from "./util/save-manager"
