+++ src/scripts/module-data/rt/entities/templates-entity.ts
@@ -1,5 +1,5 @@
-import { ColorSwatch } from '../../util/color-swatch'
-import { EntityMap } from "../../entities/create-entity"
+import { EntityMap } from "../../../entities/create-entity"
+import { ColorSwatch } from '../../../util/color-swatch'
 
 export const bigDemoEntity: EntityMap = {
     name: 'bigDemo',
