+++ src/scripts/module-data/rt/util/templates-world.ts
@@ -1,4 +1,4 @@
-import { WorldMap } from "../../util/create-world"
+import { WorldMap } from "../../../util/create-world"
 import { bigDemoEntity, demoEntity, growthDemoEntity, player } from "../entities/templates-entity"
 
 export const demoWorld: WorldMap = {
